service/pyproject.toml
[project]
name = "hello-fastapi"
version = "0.1.0"
description = "FastAPI + DDD + RBAC 权限认证系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "classy-fastapi>=0.7.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "limits>=3.10.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
]

[project.scripts]
hello-fastapi = "scripts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src"]

[tool.ruff.format]
# 跳过魔术尾随逗号，避免函数参数和装饰器参数被强制换行
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008"]

# 前端约定驼峰 JSON / Query 参数，与 PEP8 命名 intentionally 不一致
[tool.ruff.lint.per-file-ignores]
"src/api/**/*.py" = ["N803", "N815"]
"src/application/dto/**/*.py" = ["N815"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
# 与 format.skip-magic-trailing-comma 保持一致，不因尾随逗号拆分导入
split-on-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
    "ignore::DeprecationWarning:sqlalchemy",
    "ignore::ResourceWarning",
    "ignore::RuntimeWarning",
]
